requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
idna==3.4
oauthlib==3.2.2
packaging==26.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.11.0
